fix for new updates
