--- .template.config/ide.host.json
@@ -1,20 +1,0 @@
-{
-  "$schema": "http://json.schemastore.org/vs-2017.3.host",
-  "icon": "icon.png",
-  "symbolInfo": [
-    {
-      "id": "AuthorName",
-      "name": {
-        "text": "Author Name"
-      },
-      "isVisible": "true"
-    },
-    {
-      "id": "GorillaTagLocation",
-      "name": {
-        "text": "Gorilla Tag Location"
-      },
-      "isVisible": "true"
-    }
-  ]
-}
