--- .template.config/template.json
@@ -1,74 +1,0 @@
-{
-  "$schema": "http://json.schemastore.org/template",
-  "author": "Graic",
-  "classifications": ["Library"],
-  "identity": "graic.gorillatagmodtemplate.00",
-  "name": "Gorilla Tag Mod",
-  "shortName": "gtmod",
-  "tags": {
-    "type": "project",
-    "platform": "Windows",
-    "language": "C#"
-  },
-  "sourceName": "GorillaTagModTemplateProject",
-  "defaultName": "Mod",
-  "sources": [
-    {
-      "modifiers": [
-        {
-          "exclude":["*-v.zip"]
-        }
-      ]
-    }
-  ],
-  "symbols": {
-    "Framework": {
-      "type": "parameter",
-      "description": "The target framework for the project.",
-      "datatype": "choice",
-      "choices": [
-        {
-          "choice": "netstandard2.0",
-          "description": "Target netstandard2.0"
-        }
-      ],
-      "replaces": "netstandard2.0",
-      "defaultValue": "netstandard2.0"
-    },
-    "AuthorName": {
-      "type": "parameter",
-      "defaultValue": "author",
-      "replaces": "AuthorName",
-      "datatype": "text",
-      "description": "author name",
-      "isRequired": true
-    },
-    "AuthorLowerName": {
-      "type": "generated",
-      "datatype": "text",
-      "generator": "casing",
-      "parameters": {
-        "source": "AuthorName",
-        "toLower": true
-      },
-      "replaces": "AuthorLowerName"
-    },
-    "GorillaTagLocation": {
-      "type": "parameter",
-      "defaultValue": "C:\\Program Files (x86)\\Steam\\steamapps\\common\\Gorilla Tag",
-      "replaces": "C:\\Program Files (x86)\\Steam\\steamapps\\common\\Gorilla Tag",
-      "datatype": "text",
-      "description": "Gorilla Tag install location"
-    },
-    "PluginLowerName": {
-      "type": "generated",
-      "datatype": "text",
-      "generator": "casing",
-      "parameters": {
-        "source": "name",
-        "toLower": true
-      },
-      "replaces": "GorillaTagLowerModTemplateProject"
-    }
-  }
-}
