+++ HarmonyPatches.cs
@@ -2,7 +2,7 @@
 using System.Reflection;
 using HarmonyLib;
 
-namespace GorillaTagModTemplateProject
+namespace HandController
 {
 	/// <summary>
 	/// This class handles applying harmony patches to the game.
