+++ PluginInfo.cs
@@ -1,4 +1,4 @@
-﻿namespace GorillaTagModTemplateProject
+﻿namespace HandController
 {
 	/// <summary>
 	/// This class is used to provide information about your mod to BepInEx.
@@ -7,6 +7,6 @@
 	{
 		public const string GUID = "com.Deactivated.gorillatag.HandPCController";
 		public const string Name = "Hand controller";
-		public const string Version = "0.6.0";
+		public const string Version = "0.8.0";
 	}
 }
